Add fabric on Android (#402)
* Configure build.gradle

* Update libraryName on android

* Create ReactSlider shared implementation

* Split implementations into oldarch and newarch

* Dispatch events

* Cleanup eventDispatcher

* Make oldarch implementation use shared code

* Add defaults to js spec

* Clean up newarch ReactSliderManager

* Reorder props to fix disabled state

* Handle TestID setter

* Move ReactSliderShadowNode to shared implementation

* Share getExportedCustomDirectEventTypeConstants

* Remove comments, add empty line
